+++ CHANGELOG.md
@@ -1,9 +1,17 @@
+## 0.1.3
+
+- Specify audio output device.
+- Add Writer class.
+
 ## 0.1.2
+
 - Add generate_chord method.
 
 ## 0.1.1
+
 - Add Synthesizer and Player classes.
 - Generate and play constant frequency wave.
 
 ## 0.1.0
+
 - Initial version.